fe/me: always use .fum file as input for middle end
Since errors can not be written to fum file, I needed to add a `Errors.showAndExit()` after `new FrontEnd()`
