+++ src/dev/flang/fe/Module.java
@@ -42,6 +42,7 @@ import dev.flang.util.FuzionConstants;
 import dev.flang.util.HasSourcePosition;
 import dev.flang.util.SourceFile;
 
+import java.util.Collection;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.SortedMap;
@@ -458,6 +459,63 @@ public abstract class Module extends ANY
   {
     return a instanceof Feature && b instanceof Feature
      || (a instanceof LibraryFeature lf && b instanceof LibraryFeature olf && lf._libModule == olf._libModule);
+  }
+
+
+  /**
+   * allInnerAndInheritedFeatures returns a complete set of inner features, used
+   * by Clazz.layout and Clazz.hasState.
+   */
+  public Collection<AbstractFeature> allInnerAndInheritedFeatures(AbstractFeature f)
+  {
+    var d = data(f);
+    var result = d._allInnerAndInheritedFeatures;
+
+    if (result == null)
+      {
+        result = new TreeSet<>();
+        result.addAll(declaredOrInheritedFeatures(f, _dependsOn).values());
+
+        for (var p : f.inherits())
+          {
+            var cf = p.calledFeature();
+            if (CHECKS) check
+              (Errors.any() || cf != null);
+
+            if (cf != null)
+              {
+                result.addAll(allInnerAndInheritedFeatures(cf));
+              }
+          }
+        d._allInnerAndInheritedFeatures = result;
+      }
+    return result;
+  }
+
+
+  /**
+   * Find feature with given name in outer.
+   *
+   * @param outer the declaring or inheriting feature
+   */
+  public AbstractFeature lookupFeature(AbstractFeature outer, FeatureName name, AbstractFeature original)
+  {
+    if (PRECONDITIONS) require
+      (outer.state().atLeast(State.LOADING));
+
+    var result = declaredOrInheritedFeatures(outer, _dependsOn).get(name);
+
+    /* NYI: CLEANUP: can this be removed?
+     *
+     * Was feature f added to the declared features of its outer features late,
+     * i.e., after the RESOLVING_DECLARATIONS phase?  These late features are
+     * currently not added to the sets of declared or inherited features by
+     * children of their outer clazz.
+     *
+     * This is a fix for #978 but it might need to be removed when fixing #932.
+     */
+    return result == null && original instanceof Feature of && of._addedLate ? original
+                                                                             : result;
   }
 
 }
